requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai
python-dotenv
pydantic==2.4.2
requests